Add 'SameSite=None' to all the cookies
This indicates that the cookie can be set when the request is sent from
a different domain. In this case, the user is on the TDR frontend but
the cookies request is sent to the domain of the CloudFront distribution,
which is different.

Chrome actively blocks cookies that are set cross-site without this
attribute.